shadowsocks-libev

+++ scripts/build_deb.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-# Copyright 2017 Roger Shimizu <rogershimizu@gmail.com>
+# Copyright 2017-2018 Roger Shimizu <rosh@debian.org>
 #
 # This is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -49,19 +49,24 @@ exit
 apt_init() {
 	DEPS="$1"
 	DEPS_BPO="$2"
+	DEPS_EXTRA="$3"
 	if [ -n "$DEPS_BPO" ]; then
 		BPO=${OSVER}-backports
 		case "$OSID" in
 		debian)
-			REPO=http://httpredir.debian.org/debian
+			REPO=http://deb.debian.org/debian
 			;;
 		ubuntu)
 			REPO=http://archive.ubuntu.com/ubuntu
 			;;
 		esac
 		sudo sh -c "printf \"deb $REPO ${OSVER}-backports main\" > /etc/apt/sources.list.d/${OSVER}-backports.list"
+		[ -n "$DEPS_EXTRA" -a "$DEPS_EXTRA" = "sloppy" ] &&
+			sudo sh -c "printf \"\\ndeb $REPO ${OSVER}-backports-sloppy main\" >> /etc/apt/sources.list.d/${OSVER}-backports.list"
 		sudo apt-get update
 		sudo apt-get install --no-install-recommends -y -t $BPO $DEPS_BPO
+		[ -n "$DEPS_EXTRA" -a "$DEPS_EXTRA" = "sloppy" ] &&
+			sudo apt-get install --no-install-recommends -y -t ${BPO}-sloppy $DEPS_BPO
 	else
 		sudo apt-get update
 	fi
@@ -73,9 +78,12 @@ apt_clean() {
 	sudo apt-get purge -y $DEPS $DEPS_BPO debhelper \
 		libbloom-dev libcork-dev libcorkipset-dev libmbedtls-dev libsodium-dev
 	sudo apt-get purge -y libcork-build-deps libcorkipset-build-deps \
-		libbloom-build-deps libsodium-build-deps mbedtls-build-deps
-	sudo apt-get purge -y simple-obfs-build-deps shadowsocks-libev-build-deps
+		libsodium-build-deps mbedtls-build-deps
+	sudo apt-get purge -y libbloom-build-deps
+	sudo apt-get purge -y simple-obfs-build-deps
+	sudo apt-get purge -y shadowsocks-libev-build-deps
 
+if [ $BUILD_KCP -eq 1 ]; then
 	sudo apt-get purge -y golang-github-klauspost-reedsolomon-build-deps \
 		golang-github-xtaci-kcp-build-deps golang-github-xtaci-smux-build-deps \
 		kcptun-build-deps
@@ -88,6 +96,8 @@ apt_clean() {
 	sudo apt-get purge -y golang-github-urfave-cli-dev
 	sudo apt-get purge -y golang-github-pkg-errors-dev \
 		golang-github-golang-snappy-dev dh-golang
+fi
+
 	sudo apt-get autoremove -y
 }
 
@@ -98,6 +108,8 @@ gbp_build() {
 	gbp clone --pristine-tar $REPO
 	cd $PROJECT_NAME
 	[ -n "$BRANCH" ] && git checkout $BRANCH
+	[ -n "$BRANCH" -a "$BRANCH" = "trusty" ] && # try to rebase the trusty patch
+		if ! git rebase master; then git rebase --abort; git rebase debian; fi
 	[ -n "$DEPS_BPO" ] && BPO_REPO="-t ${OSVER}-backports"
 	mk-build-deps --root-cmd sudo --install --tool "apt-get -o Debug::pkgProblemResolver=yes --no-install-recommends -y $BPO_REPO"
 	rm -f ${PROJECT_NAME}-build-deps_*.deb
@@ -168,7 +180,7 @@ build_install_libmbedtls() {
 if [ $BUILD_LIB -eq 1 -o $BUILD_BIN -eq 1 ]; then
 	BRANCH=$1
 	if [ $BUILD_LIB -eq 1 ]; then
-		gbp_build https://anonscm.debian.org/git/collab-maint/mbedtls.git $BRANCH
+		gbp_build https://salsa.debian.org/debian/mbedtls.git $BRANCH
 	else
 		ls libmbed*.deb 2>&1 > /dev/null ||
 			help_lib libmbedtls
@@ -181,7 +193,29 @@ fi
 build_install_libsodium() {
 if [ $BUILD_LIB -eq 1 -o $BUILD_BIN -eq 1 ]; then
 	if [ $BUILD_LIB -eq 1 ]; then
-		dsc_build http://httpredir.debian.org/debian/pool/main/libs/libsodium/libsodium_1.0.11-2.dsc
+		git clone https://github.com/gcsideal/debian-libsodium.git libsodium
+		cd libsodium; LIBSODIUM=$(dpkg-parsechangelog --show-field Version); cd -
+		dget -ud https://deb.debian.org/debian/pool/main/libs/libsodium/libsodium_${LIBSODIUM}.dsc
+		DHVER=$(dpkg -l debhelper|grep debhelper|awk '{print $3}'|head -n1)
+		cd libsodium
+		if dpkg --compare-versions $DHVER lt 10; then
+			sed -i 's/debhelper ( >= 11)/debhelper (>= 9), dh-autoreconf/' debian/control;
+			sed -i 's/debhelper ( >= 10)/debhelper (>= 9), dh-autoreconf/' debian/control;
+			echo 9 > debian/compat;
+			dch -D unstable -l~bpo~ "Rebuild as backports"
+			git add -u;
+			git commit -m "Patch to work with ubuntu"
+		elif dpkg --compare-versions $DHVER lt 11; then
+			sed -i 's/debhelper ( >= 11)/debhelper (>= 10)/' debian/control;
+			echo 10 > debian/compat;
+			dch -D unstable -l~bpo~ "Rebuild as backports"
+			git add -u;
+			git commit -m "Patch to work with ubuntu"
+		fi
+		mk-build-deps --root-cmd sudo --install --tool "apt-get -o Debug::pkgProblemResolver=yes --no-install-recommends -y"
+		rm libsodium-build-deps_*.deb
+		gbp buildpackage -us -uc --git-ignore-branch --git-tarball-dir=.. --git-export-dir=.. --git-overlay
+		cd -
 	else
 		ls libsodium*.deb 2>&1 > /dev/null ||
 			help_lib libsodium
@@ -195,7 +229,7 @@ build_install_libbloom() {
 if [ $BUILD_LIB -eq 1 -o $BUILD_BIN -eq 1 ]; then
 	BRANCH=$1
 	if [ $BUILD_LIB -eq 1 ]; then
-		gbp_build https://anonscm.debian.org/git/collab-maint/libbloom.git $BRANCH
+		gbp_build https://salsa.debian.org/bridges-team/libbloom.git $BRANCH
 	else
 		ls libbloom-dev_*.deb libbloom1_*.deb 2>&1 > /dev/null ||
 			help_lib "libbloom-dev libbloom1"
@@ -208,12 +242,13 @@ fi
 patch_sslibev_dh9() {
 if [ $BUILD_BIN -eq 1 ]; then
 	BRANCH=$1
-	gbp clone --pristine-tar https://anonscm.debian.org/git/collab-maint/shadowsocks-libev.git
+	gbp clone --pristine-tar https://salsa.debian.org/bridges-team/shadowsocks-libev.git
 	cd shadowsocks-libev
 	[ -n "$BRANCH" ] && git checkout $BRANCH
 	sed -i 's/dh $@/dh $@ --with systemd,autoreconf/' debian/rules
 	sed -i 's/debhelper (>= 10)/debhelper (>= 9), dh-systemd, dh-autoreconf/' debian/control
 	echo 9 > debian/compat
+	dch -D unstable -l~bpo~ "Rebuild as backports"
 	git add -u
 	git commit -m "Patch to work with ubuntu trusty (14.04)"
 	cd -
@@ -224,7 +259,7 @@ fi
 build_install_sslibev() {
 if [ $BUILD_BIN -eq 1 ]; then
 	BRANCH=$1
-	gbp_build https://anonscm.debian.org/git/collab-maint/shadowsocks-libev.git $BRANCH
+	gbp_build https://salsa.debian.org/bridges-team/shadowsocks-libev.git $BRANCH
 	sudo dpkg -i shadowsocks-libev_*.deb
 	sudo apt-get install -fy
 fi
@@ -234,7 +269,7 @@ fi
 build_install_simpleobfs() {
 if [ $BUILD_BIN -eq 1 ]; then
 	BRANCH=$1
-	git_build https://anonscm.debian.org/git/collab-maint/simple-obfs.git $BRANCH
+	git_build https://salsa.debian.org/bridges-team/simple-obfs.git $BRANCH
 	sudo dpkg -i simple-obfs_*.deb
 	sudo apt-get install -fy
 fi
@@ -244,7 +279,7 @@ fi
 build_install_dhgolang() {
 if [ $BUILD_KCP -eq 1 ]; then
 	BRANCH=$1
-	gbp_build https://anonscm.debian.org/cgit/pkg-go/packages/dh-golang.git $BRANCH
+	gbp_build https://salsa.debian.org/go-team/packages/dh-golang.git $BRANCH
 	sudo dpkg -i dh-golang_*.deb
 	sudo apt-get install -fy
 fi
@@ -254,7 +289,7 @@ fi
 build_install_reedsolomondev() {
 if [ $BUILD_KCP -eq 1 ]; then
 	BRANCH=$1
-	gbp_build https://anonscm.debian.org/git/pkg-go/packages/golang-github-klauspost-reedsolomon.git $BRANCH
+	gbp_build https://salsa.debian.org/go-team/packages/golang-github-klauspost-reedsolomon.git $BRANCH
 	sudo dpkg -i golang-github-klauspost-reedsolomon-dev_*.deb
 	sudo apt-get install -fy
 fi
@@ -264,7 +299,7 @@ fi
 build_install_errorsdev() {
 if [ $BUILD_KCP -eq 1 ]; then
 	BRANCH=$1
-	gbp_build https://anonscm.debian.org/git/pkg-go/packages/golang-github-pkg-errors.git $BRANCH
+	gbp_build https://salsa.debian.org/go-team/packages/golang-github-pkg-errors.git $BRANCH
 	sudo dpkg -i golang-github-pkg-errors-dev_*.deb
 	sudo apt-get install -fy
 fi
@@ -274,10 +309,11 @@ fi
 patch_urfaveclidev_xenial() {
 if [ $BUILD_KCP -eq 1 ]; then
 	BRANCH=$1
-	gbp clone --pristine-tar https://anonscm.debian.org/git/pkg-go/packages/golang-github-urfave-cli.git
+	gbp clone --pristine-tar https://salsa.debian.org/go-team/packages/golang-github-urfave-cli.git
 	cd golang-github-urfave-cli
 	[ -n "$BRANCH" ] && git checkout $BRANCH
 	sed -i 's/golang-github-burntsushi-toml-dev/golang-toml-dev/; s/golang-gopkg-yaml.v2-dev/golang-yaml.v2-dev/' debian/control
+	dch -D unstable -l~bpo~ "Rebuild as backports"
 	git add -u
 	git commit -m "Patch to work with ubuntu xenial (16.04)"
 	cd -
@@ -288,7 +324,7 @@ fi
 build_install_urfaveclidev() {
 if [ $BUILD_KCP -eq 1 ]; then
 	BRANCH=$1
-	gbp_build https://anonscm.debian.org/git/pkg-go/packages/golang-github-urfave-cli.git $BRANCH
+	gbp_build https://salsa.debian.org/go-team/packages/golang-github-urfave-cli.git $BRANCH
 	sudo dpkg -i build-area/golang-github-urfave-cli-dev_*.deb
 	sudo apt-get install -fy
 fi
@@ -298,7 +334,7 @@ fi
 build_install_snappydev() {
 if [ $BUILD_KCP -eq 1 ]; then
 	BRANCH=$1
-	gbp_build https://anonscm.debian.org/git/pkg-go/packages/golang-github-golang-snappy.git $BRANCH
+	gbp_build https://salsa.debian.org/go-team/packages/golang-github-golang-snappy.git $BRANCH
 	sudo dpkg -i golang-github-golang-snappy-dev_*.deb
 	sudo apt-get install -fy
 fi
@@ -308,7 +344,7 @@ fi
 build_install_kcpdev() {
 if [ $BUILD_KCP -eq 1 ]; then
 	BRANCH=$1
-	gbp_build https://anonscm.debian.org/git/pkg-go/packages/golang-github-xtaci-kcp.git $BRANCH
+	gbp_build https://salsa.debian.org/go-team/packages/golang-github-xtaci-kcp.git $BRANCH
 	sudo dpkg -i golang-github-xtaci-kcp-dev_*.deb
 	sudo apt-get install -fy
 fi
@@ -318,7 +354,7 @@ fi
 build_install_smuxdev() {
 if [ $BUILD_KCP -eq 1 ]; then
 	BRANCH=$1
-	gbp_build https://anonscm.debian.org/git/pkg-go/packages/golang-github-xtaci-smux.git $BRANCH
+	gbp_build https://salsa.debian.org/go-team/packages/golang-github-xtaci-smux.git $BRANCH
 	sudo dpkg -i golang-github-xtaci-smux-dev_*.deb
 	sudo apt-get install -fy
 fi
@@ -328,7 +364,7 @@ fi
 build_install_kcptun() {
 if [ $BUILD_KCP -eq 1 ]; then
 	BRANCH=$1
-	gbp_build https://anonscm.debian.org/git/pkg-go/packages/kcptun.git $BRANCH
+	gbp_build https://salsa.debian.org/go-team/packages/kcptun.git $BRANCH
 	sudo dpkg -i kcptun_*.deb
 	sudo apt-get install -fy
 fi
@@ -369,13 +405,17 @@ esac
 
 case "$OSVER" in
 jessie)
-	BPO="debhelper libsodium-dev"
+	BPO="debhelper libbloom-dev libsodium-dev"
+	BPOEXTRA=sloppy
+	;;
+stretch)
+	BPO=libbloom-dev
 	;;
 xenial)
 	BPO=debhelper
 	;;
 esac
-apt_init "git-buildpackage pristine-tar equivs" "$BPO"
+apt_init "git-buildpackage pristine-tar equivs" "$BPO" $BPOEXTRA
 
 [ $BUILD_KCP -eq 1 ] && case "$OSVER" in
 wheezy|precise|trusty)
@@ -411,7 +451,24 @@ esac
 wheezy|precise)
 	echo Sorry, your system $OSID/$OSVER is not supported.
 	;;
-jessie|stretch|unstable|sid|zesty)
+jessie|stretch|buster|testing|unstable|sid)
+	build_install_sslibev
+	;;
+artful|bionic)
+	build_install_libbloom
+	build_install_sslibev
+	build_install_simpleobfs
+	;;
+zesty)
+	build_install_libsodium
+	build_install_libbloom
+	build_install_sslibev
+	build_install_simpleobfs
+	;;
+xenial|yakkety)
+	build_install_libcork debian
+	build_install_libcorkipset debian
+	build_install_libsodium
 	build_install_libbloom
 	build_install_sslibev
 	build_install_simpleobfs
@@ -425,13 +482,6 @@ trusty)
 	patch_sslibev_dh9
 	build_install_sslibev
 	build_install_simpleobfs trusty
-	;;
-xenial|yakkety)
-	build_install_libcork debian
-	build_install_libcorkipset debian
-	build_install_libbloom
-	build_install_sslibev
-	build_install_simpleobfs
 	;;
 *)
 	echo Your system $OSID/$OSVER is not supported yet.
